requirements.txt
torch==2.6.0
pytorch-lightning==2.6.0
opencv-python==4.11.0.86
natsort==8.4.0
SimpleITK==2.5.3
pandas==2.2.3
monai==1.5.1
monai-generative==0.2.3
gast>=0.4.0
patchify>=0.2.3
pathos==0.3.4
tqdm==4.67.1
pyyaml==6.0.2
psutil==7.0.0
lime==0.2.0.1
scipy==1.13.1
scikit-learn==1.6.1
numpy==1.26.4
transformers==4.49.0
positional-encodings==6.0.3
perceiver-pytorch==0.8.8
einops==0.8.1
nibabel==5.3.2
pydicom==2.4.4
wandb==0.16.6
#flash-attn==2.7.4.post1
